api/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
python-multipart==0.0.9
requests==2.31.0
python-dotenv==1.0.1
azure-openai==1.2.0
openai==1.20.0
pydantic==2.6.3
typing_extensions==4.10.0
